array: Add array_get_item_test  (#9)
* array: Add array_get_item_test

This test checks if array items are *false* dropped. If they aren't (try removing
`plist.false_drop = true` from the `array_get_item` function) the test panics,
because the items are being freed. It shouldn't happen and we must be able to
get items as long as the array exists.

* Update Cargo.lock